refactor(circuit): decouple attributor with saes
